Temporarily install future-mobility production deps
The server's new --settings option requires a js file from the src/js
tree which in turn requires 'js-yaml'. Due to nodejs' module resolution
algorithm, the 'js-yaml' will be searched for in src/js and above, but
not in the server directory.

This workaround can be removed once the 'js-yaml' module resolution
issue is resolved upstream.