Load exhibit-future-mobility-spy before all other UIs
The kiosk-browser has issues when several instances access the same
internal DB. The DB is locked and the camera device IDs are not properly
read (actually, new ones are created). Loading the spy a bit earlier
helped working around this issue so far.

I am aware that this is just a workaround. A proper solution would not
rely on delays. But as long as several instances of the kiosk-browser
are run simoulataneously, I can't think of any other quick fix.