Repeatedly apply touch screen mapping in background
During operation, it may be necessary to apply the touch screen mapping
again in the following cases:
- the connection to the touch panel is shaky, i.e. the touch panel
disconnects and reconnects every now and then
- the display configuration changes, i.e. displays are added, removed or
display resolutions or transformations are modified

Since it is notoriously difficult to keep track of the above without
also introducing serious race conditions into the code, the touch screen
mapping script is executed every couple of seconds.
The output of the script is logged to the console if changes of the
input configuration are detected. Note that this detection already
includes a hardly fixable race condition, which luckily does not affect
the application of the touch screen mapping itself (because it is
applied regardless).